+++ product/workflows/templates/__product-name__/service/kratos-handlers/index.ts
@@ -1,2 +1,2 @@
-export { callbacks } from "./callbacks.ts";
+export { courierCallbacks, callbacks } from "./courier-callbacks.ts";
 export { makeKratosActionHandler } from "./actions.ts";
